requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
pydub>=0.25.1
cryptography>=41.0.0
argon2-cffi>=21.3.0
typer>=0.9.0
rich>=13.0.0
piexif>=1.1.3
python-docx>=0.8.11
openpyxl>=3.1.0
pypdf>=3.0.0
